+++ packages/esm-appointments-app/src/types/index.ts
@@ -1,6 +1,78 @@
+import { OpenmrsResource } from '@openmrs/esm-framework';
+
 export enum SearchTypes {
   BASIC = 'basic',
   ADVANCED = 'advanced',
   SEARCH_RESULTS = 'search_results',
   SCHEDULED_VISITS = 'scheduled-visits',
+}
+
+interface ObsData {
+  concept: {
+    display: string;
+    uuid: string;
+  };
+  value?: string | any;
+  groupMembers?: Array<{
+    concept: { uuid: string; display: string };
+    value?: string | any;
+  }>;
+  obsDatetime: string;
+}
+
+interface Encounter {
+  diagnoses: Array<any>;
+  encounterDatetime: string;
+  encounterProviders?: Array<{ provider: { person: { display: string } } }>;
+  encounterType: { display: string; uuid: string };
+  obs: Array<ObsData>;
+  uuid: string;
+  voided: boolean;
+}
+
+export interface AppointmentsFetchResponse {
+  data: Array<Appointment>;
+}
+
+export interface Appointment {
+  appointmentKind: string;
+  appointmentNumber: string;
+  comments: string;
+  endDateTime: Date | number;
+  location: OpenmrsResource;
+  patient: fhir.Patient;
+  provider: OpenmrsResource;
+  providers: Array<OpenmrsResource>;
+  recurring: boolean;
+  service: AppointmentService;
+  startDateTime: string;
+  status: string;
+  uuid: string;
+}
+
+export interface AppointmentService {
+  appointmentServiceId: number;
+  creatorName: string;
+  description: string;
+  durationMins: string;
+  endTime: string;
+  initialAppointmentStatus: string;
+  location: OpenmrsResource;
+  maxAppointmentsLimit: number | null;
+  name: string;
+  speciality: OpenmrsResource;
+  startTime: string;
+  uuid: string;
+  serviceTypes: Array<ServiceTypes>;
+}
+
+export interface ServiceTypes {
+  duration: number;
+  name: string;
+  uuid: string;
+}
+
+interface MappedEncounter extends Omit<Encounter, 'encounterType' | 'provider'> {
+  encounterType: string;
+  provider: string;
 }
