+++ packages/esm-appointments-app/translations/en.json
@@ -1,6 +1,6 @@
 {
   "actions": "Actions",
-  "addNewAppointment": "Add new appointment",
+  "addNewAppointment": "Add new Appointment",
   "advancedSearch": "Advanced search",
   "all": "All",
   "anotherVisitType": "Start another visit type",
